tvx command to extract tipset vectors.
